feat(component): main code, all files needed, and add 2 colors on theme
This commit includes the export of the Tooltip in all files needed, creates the Tooltip component,
his stories and index.js, and Adds 2 colors on the Theme, a neutral, and a support one.

BREAKING CHANGE: this "changes" the way that you can handle the colors, giving the user 2 more
options.